Code optimization for editor-support (#1945)
* Code optimization for editor-support

* Code optimization for editor-support
